Move github package to internal directory
This prevents the new Github-specific package, meant to be used only by
Fleet, from being exposed for imports by external code.
