+++ internal/cmd/controller/gitops/reconciler/gitjob.go
@@ -13,13 +13,13 @@ import (
 	"github.com/go-logr/logr"
 	fleetapply "github.com/rancher/fleet/internal/cmd/cli/apply"
 	"github.com/rancher/fleet/internal/config"
+	fleetgithub "github.com/rancher/fleet/internal/github"
 	"github.com/rancher/fleet/internal/names"
 	"github.com/rancher/fleet/internal/ocistorage"
 	ssh "github.com/rancher/fleet/internal/ssh"
 	v1alpha1 "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	"github.com/rancher/fleet/pkg/cert"
 	fleetevent "github.com/rancher/fleet/pkg/event"
-	fleetgithub "github.com/rancher/fleet/pkg/github"
 	"github.com/rancher/fleet/pkg/sharding"
 
 	appsv1 "k8s.io/api/apps/v1"
