+++ internal/github/auth.go
